fix(vectors): compare publish-stage counts in web_source ensure_full_write; surface redaction skips; bump cli 7.1.5 (#454)
The web_source publish invariant (mark_vectors_for_completed_generation /
ensure_full_write) compared chunks_prepared (a preparation-stage count)
against points_written (a publish-stage count). When a chunk's payload
tripped the secret-redaction ForbiddenValue check and was skipped
(do-not-index-secrets), expected != wrote and the entire web-source job
failed -- e.g. the code.claude.com job wrote 7220 of 7220 attempted points
but expected 7347, the 127-point gap being exactly the redaction-skipped
chunks. The invariant now compares two publish-stage numbers
(points_attempted vs points_written), per the stage-result and
observability contracts' separation of axon_chunks_prepared_total
(preparation) from axon_vector_points_written_total (publish).

Secret-redaction chunk skips are now observable instead of silent.
Previously the only signal was a tracing::warn! line that no programmatic
consumer saw. Added VectorPointBatchBuilder::build_with_skipped_count()
returning (VectorPointBatch, skipped_count); build() delegates unchanged.
Web and non-web (git/feed/youtube/reddit/session/registry) sources surface
the skip count as a SourceWarning (*.vectorize.redaction_skipped_chunks)
on job events and the source result; local sources emit an aggregated
tracing::warn! (no SourceWarning channel on their vectorize result).

Affects CLI, MCP, and REST equally -- all three route web-URL sources
through the shared web_source pipeline; no transport divergence.

Extracted normalize_changed_documents + refetch_without_conditional +
NormalizedWebDocuments from web_source/vectorize.rs into a new
web_source/normalize.rs sibling to keep vectorize.rs under the monolith
500-line cap.

Docs: docs/pipeline-unification/runtime/observability-contract.md gains
a 'Vector point redaction skips' subsection documenting the skip behavior
and the two warning codes, plus a metrics-table note that publish-write
invariants compare publish-stage counts only.

Tests: two new tests in point_tests.rs cover build_with_skipped_count for
both the skip (forbidden chunk_text) and the no-skip paths. All 40 vectors
tests and 800 services tests pass; clippy and fmt clean.
