+++ crates/axon-services/src/web_source/vectorize.rs
@@ -1,4 +1,3 @@
-use axon_adapters::{SourceAdapter, web};
 use axon_api::source::*;
 use axon_document::{DocumentPreparer, PrepareSourceDocumentRequest};
 use axon_embedding::batch::EmbeddingBatchBuilder;
@@ -8,10 +7,8 @@ use axon_vectors::store::VectorStore;
 use uuid::Uuid;
 
 use super::WebSourceIndexInput;
-use super::artifacts::{
-    WebArtifactIndex, cleanup_artifacts_after_error, store_clean_outputs, store_warc_artifact,
-};
-use super::reuse;
+use super::artifacts::{WebArtifactIndex, cleanup_artifacts_after_error};
+use super::normalize::{NormalizedWebDocuments, normalize_changed_documents};
 use super::run::{WebAdapterRun, timestamp};
 use super::vectorize_helpers::{
     changed_diff_batches, document_status, payload_index, prepared_document_batches,
@@ -25,6 +22,9 @@ const WEB_CHANGED_CHUNK_BATCH_SIZE: usize = 512;
 pub(super) struct VectorizeResult {
     pub(super) documents_prepared: u64,
     pub(super) chunks_prepared: u64,
+    /// Publish-stage count: vector points actually upserted (post-skip). Fed
+    /// to the publish invariant instead of `chunks_prepared`.
+    pub(super) points_attempted: u64,
     pub(super) document_statuses: Vec<DocumentStatus>,
     pub(super) reused_item_keys: Vec<SourceItemKey>,
     /// Parser-produced graph candidates carried by each prepared document
@@ -43,6 +43,21 @@ pub(super) struct VectorizeResult {
 struct VectorizeStats {
     documents_prepared: u64,
     chunks_prepared: u64,
+    /// Publish-stage count: vector points actually upserted (the post-skip
+    /// value). Distinct from `chunks_prepared`, which is the preparation-stage
+    /// count of emitted chunks before secret-redaction skips. The publish
+    /// invariant (`ensure_full_write`) compares against this, not
+    /// `chunks_prepared`, so that chunks skipped by the redaction
+    /// `ForbiddenValue` check don't make expected != wrote. See
+    /// `docs/pipeline-unification/runtime/observability-contract.md`'s
+    /// `axon_chunks_prepared_total` vs `axon_vector_points_written_total`.
+    points_attempted: u64,
+    /// Chunks skipped during point-batch build because their payload tripped
+    /// the secret-redaction `ForbiddenValue` check (do-not-index-secrets).
+    /// Surfaced as a `SourceWarning` on the `VectorizeResult` so it is
+    /// observable to job events and the source result, not only as a
+    /// `tracing::warn!`.
+    chunks_skipped_redaction: u64,
 }
 
 #[derive(Debug, Clone, Default)]
@@ -51,15 +66,6 @@ struct VectorizeResultWithStats {
     document_statuses: Vec<DocumentStatus>,
     graph_candidates: Vec<GraphCandidate>,
     warnings: Vec<SourceWarning>,
-}
-
-pub(super) struct NormalizedWebDocuments {
-    pub(super) documents: Vec<SourceDocument>,
-    pub(super) warnings: Vec<SourceWarning>,
-    pub(super) reused_item_keys: Vec<SourceItemKey>,
-    pub(super) artifacts: Vec<ArtifactRef>,
-    pub(super) inline: Option<InlineSourceResult>,
-    pub(super) artifact_index: WebArtifactIndex,
 }
 
 pub(super) async fn vectorize_changed_documents(
@@ -116,6 +122,20 @@ pub(super) async fn vectorize_changed_documents(
             };
             result.documents_prepared += batch_result.stats.documents_prepared;
             result.chunks_prepared += batch_result.stats.chunks_prepared;
+            result.points_attempted += batch_result.stats.points_attempted;
+            if batch_result.stats.chunks_skipped_redaction > 0 {
+                result.warnings.push(SourceWarning {
+                    code: "web.vectorize.redaction_skipped_chunks".to_string(),
+                    severity: Severity::Warning,
+                    message: format!(
+                        "skipped {} chunk(s) with secret-redaction-forbidden payload values \
+                         (not indexed; reduced vector point count accordingly)",
+                        batch_result.stats.chunks_skipped_redaction
+                    ),
+                    source_item_key: None,
+                    retryable: false,
+                });
+            }
             result
                 .document_statuses
                 .extend(batch_result.document_statuses);
@@ -216,189 +236,6 @@ pub(super) fn published_status(status: &DocumentStatus) -> DocumentStatus {
     }
 }
 
-pub(super) async fn normalize_changed_documents(
-    input: &WebSourceIndexInput,
-    run: &WebAdapterRun,
-    diff: &SourceManifestDiff,
-) -> anyhow::Result<NormalizedWebDocuments> {
-    let adapter = web::WebSourceAdapter::new(
-        std::sync::Arc::clone(&input.fetch_provider),
-        std::sync::Arc::clone(&input.render_provider),
-    );
-    let mut acquisition = adapter.acquire(&run.plan, diff).await?;
-    let mut warnings = acquisition.header.warnings.clone();
-    let mut artifact_index = WebArtifactIndex::default();
-    let mut artifacts = Vec::new();
-    for artifact in store_warc_artifact(input, run, &acquisition.fetched_items).await? {
-        artifact_index.push_generation(artifact.clone());
-        artifacts.push(artifact);
-    }
-    let mut documents = Vec::new();
-    let mut documents_to_cache = Vec::new();
-    let mut fetched_items = Vec::new();
-    let mut reused_item_keys = Vec::new();
-
-    for item in std::mem::take(&mut acquisition.fetched_items) {
-        let reuse_required = item
-            .metadata
-            .get("web_reuse_required")
-            .and_then(serde_json::Value::as_bool)
-            .unwrap_or(false);
-        if !reuse_required {
-            fetched_items.push(item);
-            continue;
-        }
-
-        if let Some(reused) = reuse::load_reused_web_document(
-            input.document_cache.as_ref(),
-            &run.source_id,
-            diff.previous_generation.as_ref(),
-            &item.manifest_item.source_item_key,
-            &diff.next_generation,
-        )
-        .await?
-        {
-            reused_item_keys.push(item.manifest_item.source_item_key.clone());
-            documents_to_cache.push(reused.document);
-            continue;
-        }
-
-        warnings.push(SourceWarning {
-            code: "web.reuse.cache_miss_refetch".to_string(),
-            severity: Severity::Warning,
-            message: format!(
-                "conditional 304 for {} had no cached committed document; refetching before publish",
-                item.manifest_item.canonical_uri
-            ),
-            source_item_key: Some(item.manifest_item.source_item_key.clone()),
-            retryable: true,
-        });
-        fetched_items
-            .push(refetch_without_conditional(input, run, diff, item.manifest_item).await?);
-    }
-
-    if !fetched_items.is_empty() {
-        acquisition.fetched_items = fetched_items;
-        let normalized = adapter.normalize(&run.plan, acquisition).await?.data;
-        let clean_output = store_clean_outputs(input, &normalized).await?;
-        artifacts.extend(clean_output.artifacts);
-        artifact_index.merge(clean_output.artifact_index);
-        let inline = clean_output.inline;
-        documents_to_cache.extend(normalized.clone());
-        documents.extend(normalized);
-        reuse::cache_documents(
-            input.document_cache.as_ref(),
-            &run.source_id,
-            &diff.next_generation,
-            &documents_to_cache,
-        )
-        .await?;
-        return Ok(NormalizedWebDocuments {
-            documents,
-            warnings,
-            reused_item_keys,
-            artifacts,
-            inline,
-            artifact_index,
-        });
-    }
-
-    reuse::cache_documents(
-        input.document_cache.as_ref(),
-        &run.source_id,
-        &diff.next_generation,
-        &documents_to_cache,
-    )
-    .await?;
-    Ok(NormalizedWebDocuments {
-        documents,
-        warnings,
-        reused_item_keys,
-        artifacts,
-        inline: None,
-        artifact_index,
-    })
-}
-
-async fn refetch_without_conditional(
-    input: &WebSourceIndexInput,
-    run: &WebAdapterRun,
-    diff: &SourceManifestDiff,
-    manifest_item: ManifestItem,
-) -> anyhow::Result<AcquiredSourceItem> {
-    let mut plan = run.plan.clone();
-    plan.route
-        .validated_options
-        .values
-        .insert("etag_conditional".to_string(), serde_json::json!(false));
-    let adapter = web::WebSourceAdapter::new(
-        std::sync::Arc::clone(&input.fetch_provider),
-        std::sync::Arc::clone(&input.render_provider),
-    );
-    let reacquired = adapter
-        .acquire(
-            &plan,
-            &SourceManifestDiff {
-                header: diff.header.clone(),
-                source_id: diff.source_id.clone(),
-                previous_generation: diff.previous_generation.clone(),
-                next_generation: diff.next_generation.clone(),
-                added: Vec::new(),
-                modified: vec![manifest_item.clone()],
-                removed: Vec::new(),
-                unchanged: Vec::new(),
-                skipped: Vec::new(),
-                failed: Vec::new(),
-                counts: DiffCounts {
-                    added: 0,
-                    modified: 1,
-                    removed: 0,
-                    unchanged: 0,
-                    skipped: 0,
-                    failed: 0,
-                },
-            },
-        )
-        .await?;
-    let mut reacquired_items = reacquired.fetched_items.into_iter();
-    let reacquired = match reacquired_items.next() {
-        Some(item) => item,
-        None => {
-            if let Some(warning) = reacquired.header.warnings.iter().find(|warning| {
-                warning.code == "web.fetch.invalid_304_without_validator"
-                    || warning.message.contains("304 Not Modified")
-            }) {
-                anyhow::bail!(
-                    "unconditional refetch for {} received another 304/reuse response: {}",
-                    manifest_item.canonical_uri,
-                    warning.message
-                );
-            }
-            anyhow::bail!(
-                "unconditional refetch for {} returned no document",
-                manifest_item.canonical_uri
-            );
-        }
-    };
-    let reuse_required = reacquired
-        .metadata
-        .get("web_reuse_required")
-        .and_then(serde_json::Value::as_bool)
-        .unwrap_or(false);
-    if reuse_required
-        || matches!(
-            &reacquired.content_ref,
-            ContentRef::External { uri, .. } if uri.starts_with("reuse://")
-        )
-    {
-        anyhow::bail!(
-            "unconditional refetch for {} returned 304/reuse instead of content",
-            manifest_item.canonical_uri
-        );
-    }
-    Ok(reacquired)
-}
-
 fn prepare_source_documents(
     source_documents: Vec<SourceDocument>,
     generation: &SourceGenerationId,
@@ -439,7 +276,8 @@ async fn vectorize_documents(
     }
     let batch = embedding_batch_for_documents(input, &documents)?;
     let embeddings = embedding_provider.embed(batch).await?;
-    let point_batch = vector_point_batch_for_documents(collection, &documents, &embeddings)?;
+    let (point_batch, skipped_redaction) =
+        vector_point_batch_for_documents(collection, &documents, &embeddings)?;
     let expected_points = point_batch.points.len() as u64;
     let write = vector_store.upsert(point_batch).await?;
     if write.points_attempted != write.points_written || write.points_written != expected_points {
@@ -450,6 +288,8 @@ async fn vectorize_documents(
         ));
     }
     let mut result = VectorizeResultWithStats::default();
+    result.stats.points_attempted = write.points_attempted;
+    result.stats.chunks_skipped_redaction = skipped_redaction;
     for document in documents {
         result.stats.chunks_prepared += document.chunks.len() as u64;
         result.stats.documents_prepared += 1;
